Desktop/Claude_Projects/CH_HiScore/requirements.txt
# Clone Hero High Score System - Python Dependencies

# File monitoring
watchdog>=3.0.0

# Discord bot
discord.py>=2.3.0
python-dotenv>=1.0.0

# HTTP API
aiohttp>=3.9.0
requests>=2.31.0

# OCR for results screen capture (uses Windows built-in OCR)
winocr>=0.0.15
mss>=9.0.0
Pillow>=10.0.0
pywin32>=306

# System tray support (optional)
pystray>=0.19.0

# Database (built-in sqlite3 is sufficient for now)

# Packaging (for creating executable)
pyinstaller>=6.0.0
